refactor: simplify project route by replacing Prisma queries with utility functions for improved readability and maintainability; update export options in the frontend to ensure pecha template format is correctly integrated
