+++ backend/utils/const.js
@@ -1,0 +1,7 @@
+const path = require("path");
+
+const TEMPLATE_PATH = path.join(__dirname, "..", "static", "template.docx");
+
+module.exports = {
+  TEMPLATE_PATH,
+};
